requirements.txt
matplotlib==3.2.2
natsort==8.2.0
numpy==1.18.5
Pillow==9.3.0
PyYAML==6.0
scikit_learn==1.2.0
torch==1.10.0
torchvision==0.11.1
tqdm==4.47.0
